refactor: Extract generate tags on csv to a function api
